基本参数维护

+++ distribution-parent/distribution-data/src/main/java/com/chauncy/data/domain/po/user/PmMemberLevelPo.java
@@ -58,7 +58,7 @@ public class PmMemberLevelPo implements Serializable {
     private BigDecimal purchasePresent;
 
     @ApiModelProperty(value = "会员等级经验值")
-    private Integer levelExperience;
+    private BigDecimal levelExperience;
 
     @ApiModelProperty(value = "红包赠送比例")
     private BigDecimal packetPresent;
